Extract ActionsMenu into separate component
