+++ packages/frontend/editor-ui/src/app/components/MainHeader/WorkflowDetails.vue
@@ -6,6 +6,7 @@ import WorkflowProductionChecklist from '@/components/WorkflowProductionChecklis
 import WorkflowTagsContainer from '@/components/WorkflowTagsContainer.vue';
 import WorkflowTagsDropdown from '@/components/WorkflowTagsDropdown.vue';
 import {
+	DRAFT_PUBLISH_EXPERIMENT,
 	MAX_WORKFLOW_NAME_LENGTH,
 	MODAL_CONFIRM,
 	PLACEHOLDER_EMPTY_WORKFLOW_ID,
@@ -13,6 +14,7 @@ import {
 } from '@/app/constants';
 
 import { useProjectsStore } from '@/features/collaboration/projects/projects.store';
+import { usePostHog } from '@/stores/posthog.store';
 import { useDocumentTitle } from '@/composables/useDocumentTitle';
 import { useMessage } from '@/composables/useMessage';
 import { useTelemetry } from '@/composables/useTelemetry';
@@ -26,6 +28,7 @@ import { useNpsSurveyStore } from '@/app/stores/npsSurvey.store';
 import { ProjectTypes } from '@/features/collaboration/projects/projects.types';
 import type { PathItem } from '@n8n/design-system/components/N8nBreadcrumbs/Breadcrumbs.vue';
 import WorkflowHeaderActions from '@/components/MainHeader/WorkflowHeaderActions.vue';
+import WorkflowHeaderDraftPublishActions from '@/components/MainHeader/WorkflowHeaderDraftPublishActions.vue';
 import { useI18n } from '@n8n/i18n';
 import { getResourcePermissions } from '@n8n/permissions';
 import { createEventBus } from '@n8n/utils/event-bus';
@@ -80,6 +83,7 @@ const projectsStore = useProjectsStore();
 const foldersStore = useFoldersStore();
 const npsSurveyStore = useNpsSurveyStore();
 const i18n = useI18n();
+const posthogStore = usePostHog();
 
 const router = useRouter();
 const route = useRoute();
@@ -138,6 +142,10 @@ const currentFolderForBreadcrumbs = computed(() => {
 		return foldersStore.getCachedFolder(folderId);
 	}
 	return null;
+});
+
+const isDraftPublishEnabled = computed(() => {
+	return posthogStore.isFeatureEnabled(DRAFT_PUBLISH_EXPERIMENT.name);
 });
 
 watch(
@@ -565,7 +573,9 @@ onBeforeUnmount(() => {
 
 		<PushConnectionTracker class="actions">
 			<WorkflowProductionChecklist v-if="!isNewWorkflow" :workflow="workflowsStore.workflow" />
+			<WorkflowHeaderDraftPublishActions v-if="isDraftPublishEnabled" />
 			<WorkflowHeaderActions
+				v-else
 				:id="id"
 				ref="workflowHeaderActions"
 				:name="name"
