+++ packages/frontend/editor-ui/src/app/components/MainHeader/WorkflowHeaderDraftPublishActions.vue
@@ -1,0 +1,5 @@
+<script lang="ts" setup></script>
+
+<template>
+	<div></div>
+</template>
